requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
idna==2.9
Naked==0.1.31
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.9.7
pypiwin32==223
pywin32==227
PyYAML==5.3
requests==2.23.0
rsa==4.0
shellescape==3.8.1
six==1.14.0
urllib3==1.25.8
